- Prepare for Release
- updated gradle to 4.4
- update kotlin to 1.2.30
- updated package structure
- added travis yml file
