+++ .travis.yml
@@ -1,0 +1,7 @@
+language: java
+jdk:
+- oraclejdk8
+before_install:
+- chmod +x gradlew
+script:
+- "./gradlew clean build"
