+++ src/main/kotlin/org/rekotlin/Action.kt
@@ -1,4 +1,4 @@
-package tw.geothings.rekotlin
+package org.rekotlin
 
 /**
  * Created by Taras Vozniuk on 31/07/2017.
@@ -25,24 +25,13 @@ package tw.geothings.rekotlin
  */
 
 /**
- * Defines the interface of a dispatching, stateless Store in ReSwift. `StoreType` is
- * the default usage of this interface. Can be used for store variables where you don't
- * care about the state, but want to be able to dispatch actions.
+ * All actions that want to be able to be dispatched to a store need to conform to this protocol
+ * Currently it is just a marker protocol with no requirements.
  */
-interface DispatchingStoreType {
+interface Action {}
 
-    /**
-     * Dispatches an action. This is the simplest way to modify the stores state.
-     *
-     * Example of dispatching an action:
-     * <pre>
-     * <code>
-     * store.dispatch( CounterAction.IncreaseCounter )
-     * </code>
-     * </pre>
-     *
-     * @param action The action that is being dispatched to the store
-     * @return By default returns the dispatched action, but middlewares can change the type, e.g. to return promises
-     */
-    fun dispatch(action: Action)
-}
+/**
+ * Initial Action that is dispatched as soon as the store is created.
+ * Reducers respond to this action by configuring their initial state.
+ */
+class ReKotlinInit: Action {}
