+++ src/main/kotlin/org/rekotlin/State.kt
@@ -1,4 +1,4 @@
-package tw.geothings.rekotlin
+package org.rekotlin
 
 /**
  * Created by Taras Vozniuk on 31/07/2017.
@@ -24,14 +24,5 @@ package tw.geothings.rekotlin
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-/**
- * All actions that want to be able to be dispatched to a store need to conform to this protocol
- * Currently it is just a marker protocol with no requirements.
- */
-interface Action {}
+interface StateType {}
 
-/**
- * Initial Action that is dispatched as soon as the store is created.
- * Reducers respond to this action by configuring their initial state.
- */
-class ReKotlinInit: Action {}
