+++ src/main/kotlin/org/rekotlin/Subscription.kt
@@ -1,4 +1,4 @@
-package tw.geothings.rekotlin
+package org.rekotlin
 
 /**
  * Created by Taras Vozniuk on 07/08/2017.
@@ -36,8 +36,8 @@ package tw.geothings.rekotlin
  *
  */
 class SubscriptionBox<State, SelectedState>(val originalSubscription: Subscription<State>,
-                             transformedSubscription: Subscription<SelectedState>?,
-                             val subscriber: StoreSubscriber<SelectedState>) where State: StateType {
+                                            transformedSubscription: Subscription<SelectedState>?,
+                                            val subscriber: StoreSubscriber<SelectedState>) where State: StateType {
 
     // hoping to mimic swift weak reference
     // however this doesn't really work the same way, gc collects non-deterministically
@@ -109,11 +109,11 @@ class Subscription<State> {
      * thus should be skipped and not forwarded to subscribers.
      *
      */
-    fun skipRepeats(isRepeat: (oldState: State, newState: State) -> Boolean): Subscription<State>{
+    fun skipRepeats(isRepeat: (oldState: State, newState: State) -> Boolean): Subscription<State> {
         return Subscription { sink ->
             this.observe { oldState, newState ->
                 oldState?.let {
-                    if (!isRepeat(oldState, newState)){
+                    if (!isRepeat(oldState, newState)) {
                         sink(oldState, newState)
                     }
 
@@ -126,7 +126,7 @@ class Subscription<State> {
      * Provides a subscription that skips repeated updates of the original subscription
      * Repeated updates determined by structural equality
      */
-    fun skipRepeats(): Subscription<State>{
+    fun skipRepeats(): Subscription<State> {
         return this.skipRepeats { oldState, newState ->
             oldState == newState
         }
@@ -138,7 +138,7 @@ class Subscription<State> {
      * @param when A closure that determines whether a given state update is a repeat and
      * thus should be skipped and not forwarded to subscribers.
      */
-    fun skip(`when`: (oldState: State, newState: State) -> Boolean): Subscription<State>{
+    fun skip(`when`: (oldState: State, newState: State) -> Boolean): Subscription<State> {
         return this.skipRepeats(`when`)
     }
 
@@ -148,7 +148,7 @@ class Subscription<State> {
      * This is effectively the inverse of skipRepeats(:)
      * @param whenBlock A closure that determines whether a given state update should notify
      */
-    fun only(whenBlock: (oldState: State, newState: State) -> Boolean): Subscription<State>{
+    fun only(whenBlock: (oldState: State, newState: State) -> Boolean): Subscription<State> {
         return this.skipRepeats { oldState, newState ->
             !whenBlock(oldState, newState)
         }
